# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Folio Expense Tracker API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "CatalogEntry": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "categoryId": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "manufacturerId": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          }
        }
      },
      "InvoiceInput": {
        "type": "object",
        "required": [
          "date",
          "shopId",
          "currency",
          "lines"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "shopId": {
            "type": "integer",
            "minimum": 1
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "KES",
              "GBP",
              "CHF",
              "CAD",
              "AUD"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "itemId",
                "quantity",
                "unitPrice"
              ],
              "properties": {
                "itemId": {
                  "type": "integer",
                  "minimum": 1
                },
                "quantity": {
                  "type": "number",
                  "exclusiveMinimum": true,
                  "minimum": 0,
                  "maximum": 1000,
                  "multipleOf": 0.001
                },
                "unitPrice": {
                  "type": "string",
                  "pattern": "^\\d{1,7}(\\.\\d{1,2})?$",
                  "example": "12.50"
                }
              }
            }
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "shopId": {
            "type": "integer"
          },
          "shop": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "totalCents": {
            "type": "integer",
            "nullable": true
          },
          "unpriced": {
            "type": "boolean"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "itemId": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "quantity": {
                  "type": "number"
                },
                "unitPriceCents": {
                  "type": "integer",
                  "nullable": true
                }
              }
            }
          }
        }
      },
      "MonthlySnapshot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "month": {
            "type": "string",
            "example": "2026-08"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ready"
            ]
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "cutoffDate": {
            "type": "string",
            "format": "date"
          },
          "timeZone": {
            "type": "string"
          },
          "catchUp": {
            "type": "boolean"
          },
          "invoiceCount": {
            "type": "integer"
          },
          "pricedCount": {
            "type": "integer"
          },
          "unpricedCount": {
            "type": "integer"
          },
          "originalTotals": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Original minor-unit totals keyed by currency."
          },
          "totals": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Same combined spending expressed separately in each supported currency, in minor units."
          },
          "rates": {
            "type": "object",
            "nullable": true,
            "description": "Provider, URL, requested date, base EUR, rates and individual publication dates."
          },
          "invoices": {
            "type": "array",
            "description": "Frozen invoice details; returned by detail endpoint only.",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          },
          "kind": {
            "type": "string",
            "enum": [
              "monthly",
              "ad_hoc"
            ]
          },
          "title": {
            "type": "string"
          },
          "fromDate": {
            "type": "string",
            "format": "date"
          },
          "toDate": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Sign in; sets a two-hour HttpOnly session cookie",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          },
          "429": {
            "description": "Too many login attempts"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "security": []
      }
    },
    "/logout": {
      "post": {
        "summary": "Clear session cookie",
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": []
      }
    },
    "/me": {
      "get": {
        "summary": "Return id, username, name, canManage and defaultCurrency",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/catalog": {
      "get": {
        "summary": "List shared shops, items, categories and manufacturers",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "description": "Each shop, item, category and manufacturer exposes userId, the creator ID or null for legacy entries. Catalog remains shared."
      }
    },
    "/shops": {
      "post": {
        "summary": "Create catalog entry; requires userrole_manageshops",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            }
          }
        },
        "description": "Stores user_id from the authenticated user. Client-supplied user IDs are ignored. Returns id and userId."
      }
    },
    "/items": {
      "post": {
        "summary": "Create catalog entry; requires userrole_manageshops",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            }
          }
        },
        "description": "Stores user_id from the authenticated user. Client-supplied user IDs are ignored. Returns id and userId."
      }
    },
    "/categories": {
      "post": {
        "summary": "Create catalog entry; requires userrole_manageshops",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            }
          }
        },
        "description": "Stores user_id from the authenticated user. Client-supplied user IDs are ignored. Returns id and userId."
      }
    },
    "/manufacturers": {
      "post": {
        "summary": "Create catalog entry; requires userrole_manageshops",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            }
          }
        },
        "description": "Stores user_id from the authenticated user. Client-supplied user IDs are ignored. Returns id and userId."
      }
    },
    "/invoices": {
      "get": {
        "summary": "List only the signed-in user's invoices; unpriced legacy totals are null",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Create an expense with its lines atomically",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceInput"
              }
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "put": {
        "summary": "Replace your expense and its lines atomically",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          },
          "404": {
            "description": "Expense absent or owned by another user"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceInput"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete your expense and its lines atomically",
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin or permission denied"
          },
          "500": {
            "description": "Internal error"
          },
          "404": {
            "description": "Expense absent or owned by another user"
          }
        }
      }
    },
    "/snapshots": {
      "get": {
        "summary": "List the authenticated user's monthly and ad hoc summaries, newest capture first",
        "responses": {
          "200": {
            "description": "Snapshots without invoice details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthlySnapshot"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sign in required"
          }
        }
      },
      "post": {
        "summary": "Create a private immutable ad hoc summary immediately",
        "description": "Captures an inclusive date range, uses end-date reference rates, and does not change automatic monthly summaries. End date cannot exceed today in the configured snapshot timezone. Returns ready or pending if exchange rates are unavailable.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "from",
                  "to"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "from": {
                    "type": "string",
                    "format": "date"
                  },
                  "to": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saved snapshot with frozen details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlySnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date range or title"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/snapshots/{id}": {
      "get": {
        "summary": "Read a frozen monthly picture belonging to the authenticated user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot with original invoice details and saved conversions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlySnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id"
          },
          "401": {
            "description": "Sign in required"
          },
          "404": {
            "description": "Not found or belongs to another user"
          }
        }
      }
    }
  }
}